continue applying best practices for RESTful interfaces
Each stored result now has an individual endpoint under
"/benchmark-results/".  GET and DELETE methods can now be used to retrieve
or remove a result correspondingly.
PUT and PATCH methods are not supported.

While there, rename the submit method of BenchmarkAPI_V1 to "post", so
that it matches the naming convention used for other methods.

test_submit_persists() is changed to use the GET HTTP request instead
of directly peeking into the backend.
Also, tests for the DELETE request are added.