add a "real" backend that uses txmongo driver to work with MongoDB
Note the following limitations:
- txmongo does not support 'sort' parameter for collection.find() method
  and the parameter is silently ignored, so query() method of the backend
  is broken now
- currently the backend relies on auto-generated IDs being monotonically
  increasing, but that's no guaranteed
- moreover, we should use a timestamp embedded into documents for sorting

Also, it's not clear if we should include the "_id" field with the
retrieved documents.  If we decide to do so, then we need to take care
to properly serialize its ObjectId value (as a string, probably).
Also, it would make sense to return the full document including the _id
field as a response to a POST that creates a new document.

Note that at the moment the new backend is not used and thus not tested.