re-implement the query interface as a GET method
Currently this interface supports filtering only by the branch name.
Note that the code expects that the branch name is provided in the result
content as "userdata":"branch".
There is no support for the paging of the results,
but a limit on the number of the returned items is supported.
The order of the results is fixed at the moment: from the latest submitted
to the earliest.