first argument of txmongo's find_one() is different from 'filter' argument
It seems that the first argument that actually specifies the query is
named 'spec' in txmongo and the 'filter' argument does something different
(not sure what).

Previously the code passed the tests because the testing was done
essentially against an empty collection, so there was no chance of
acting on a wrong record.  Now the tests have been updated to have
an unrelated record.